Reorganize noop function
