--- src/utils/internal.ts
@@ -1,5 +1,0 @@
-/**
- * Utility functions for internal use only.
- */
-
-export function noop() {}
