+++ src/utils/noop.ts
@@ -1,0 +1,4 @@
+/**
+ * No-op function to prevent lint errors.
+ */
+export default function noop() {}
